+++ status_bar.h
@@ -35,8 +35,8 @@ public:
         solidroundrect(position.x + 100, position.y + 10, position.x + 100 + width + 3, position.y + 33, 8, 8);
         solidroundrect(position.x + 100, position.y + 45, position.x + 100 + width + 3, position.y + 68, 8, 8);
 
-        float hp_bar_width = width * max(0, hp) / 100.0f;
-        float mp_bar_width = width * min(100, mp) / 100.0f;
+        float hp_bar_width = (float)width * max(0, hp) / 100.0f;
+        float mp_bar_width = (float)width * min(100, mp) / 100.0f;
         setfillcolor(RGB(197, 61, 67));
         solidroundrect(position.x + 100, position.y + 10, position.x + 100 + (int)hp_bar_width + 3, position.y + 33, 8, 8);
         setfillcolor(RGB(83, 131, 195));
